pyproject.toml
[project]
name = "vival"
version = "0.1.0"
description = "A visual AI dictionary"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2>=2.9.10",
    "openai>=1.55.1",
    "python-dotenv>=1.0.1",
    "dash>=2.18.2",
    "dash-mantine-components>=0.15.1",
    "dash-iconify>=0.1.2",
    "diskcache>=5.6.3",
    "pillow>=11.1.0",
    "ipykernel>=6.29.5",
]


[tool.ruff]
line-length = 120

[tool.setuptools]
packages = { find = { where = ["src"] } }

[tool.uv]
cache-dir = "/workspace/.cache"